requirements.txt
arrow==1.2.0
asgiref==3.4.1
autopep8==1.5.7
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
dj-database-url==0.5.0
Django==3.2.5
django-activeurl==0.2.0
django-appconf==1.0.5
django-classy-tags==2.0.0
django-heroku==0.3.1
django-tailwind==2.2.2
gunicorn==20.1.0
idna==2.10
Jinja2==3.0.2
jinja2-time==0.2.0
lxml==4.7.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
poyo==0.5.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
python-decouple==3.5
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
